perform runtime check only for non-collection types
Since collection types are a special case when comparing props,
line 38 already provides an accurate runtime check to ensure that both
units are collection types. This allows the collection library to
properly evaluate deep equality on the provided collection. If the
runtimeType is compared before this, it's possible to get a false
negative. List, for example, is an abstract class and there are many
possible implementations. Each of these implementations would be a List
but their runtimeType is different which would prevent us from
comparing the two lists. For this reason, we should check for the
collection special case first, and then use the runtimeType check for
all other types.