Update README.md to use List<Object?>
I tried to add Equatable to a class I have according to the readme, and it failed because I had a nullable field.
Replacing `List<Object>` with `List<Object?>` fixed the problem. So I suggest to change the readme.

The actual overridden property is indeed `List<Object?> get props`, so I think there's no need to restrict it to non-nullable values.